"Implement player movement and property ownership"
This commit expands on the functionality of player movement and property ownership in our game. We refined our Property model, adding an owner attribute and buy method which enables property acquisition by the players.

The Game model was updated to include a main game loop, and we added change_location method and location attribute to our Player model. We created a new utility function to fetch the names of properties owned by player.

Finally, updates to the main game start scripts were made in the main.py file to reflect these changes, and we improved code organization by renaming and moving the property_collectors.py file to the Utils directory. These changes provide the core gameplay loop necessary for our game.