Update rent calculation and game flow
This revision implements changes to the rent calculation function 'determineRent' in buying.py to include the player object. This fixes a bug where a player was deducted rent even if they were the owner of the location.

In game.py, the changes eliminate the use of private variables for 'players' and 'squares'. The update also includes logic for handling bankruptcy and determining the winner of the game. These changes enhance the gameplay and provide clear end-game parameters.