Added chance and community chest card data, moved square data
Created new data files 'chance.txt' and 'comchest.txt' to hold information for chance and community chest cards in the game. The information includes the name of the card, the type of card, and values associated with the card.

Also, moved square data to a new Data directory for better file organization. Now all game-relevant data is stored within the Data directory. This change enhances the readability and structure of the project, making it easier to locate and manage the various game data.