Add ability to build houses on properties and update rent calculation
This update adds a method to build houses on owned properties, so long as the player owns all properties of the same color. This checks the ownership and color of the properties, as well as the player's balance to ensure they can afford the house. This feature opens up new strategic possibilities for players.

Additional updates were done on the method to calculate rent. For utilities, the rent is calculated depending on the roll size and the number of utilities owned, adding another layer of complexity to the game logic.

Added case handling for jail situations in player model, preparing for future feature releases. Jail related attributes and methods were added to handle such scenarios.

Input check has been improved in buying properties, transforming user inputs to lowercase, thus eliminating potential errors due to casing. This will better handle varying user inputs and protect against errors.

The changes improve gameplay complexity, user experience, and prepares the base model for future enhancements.