Refactor buying system to distinct move and purchase methods
Renamed 'buying.py' to 'moving.py' to better represent its responsibility for player's movement around the board. Added 'check_movement' function to handle the logic of player movement, which includes checking if a property is owned or under mortgage. Also created a new 'buy_property' function that handles proposed acquisitions and checks if players have enough balance. Restricted house building action to Streets in 'properties.py'. Initialized all properties with owner set to None in 'property_collection.py'. The function known as 'start' in 'main.py' was renamed to 'setup' for clarity.

This refactoring brings more clarity and modularity to the codebase, making it easier to maintain, debug, and extend in future developments.