Refactor property buying and movement checking code
The purpose of this commit is to make changes in property buying and movement checking mechanisms. Now, the 'buy' function from the Property model has been moved to the 'properties' utility and redesigned to include auction feature. That allows other players to bid for the property if initial player decides not to buy it. Also, an additional layer of checking whether the player can buy a property has been established - if a player's balance is insufficient, he or she will not be given an option to buy. The 'check_movement' function has been altered to accept 'all_players' argument to facilitate the auction process. These changes are expected to enhance game fairness and interaction between players.