Remove deprecated game options notes in Game class
The detailed game options notes present in the 'Game' class have been removed. This step was taken as all the options have been successfully integrated into the game. The 'Game' class now is less cluttered with the elimination of the extensive comment block that listed the functionalities yet to be incorporated in the game. The readability of the code is thus improved without compromising on its functionality.